restructure proc-macro crate to drop cfg_attr
