+++ rust/opendp_derive/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "opendp_derive"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+proc-macro = true
+
+[dependencies]
+syn = { version = "1.0.81", features = ["full", "parsing"], optional = true }
+quote = { version = "1.0.10", optional = true }
+proc-macro2 = { version = "1.0.32", optional = true }
+
+[features]
+ffi = ["syn", "quote", "proc-macro2"]
